+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
 * Added support for declaring wildcard audits and trusted entries for crates published using ["Trusted Publishing"](https://crates.io/docs/trusted-publishing) (#671)
+* `cargo vet check --frozen` (a disabled network) will infer the existence of crates using versions
+  in audits when checking audit-as-crates-io policies (#661)
 
 # Version 0.10.1 (2025-02-10)